backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
meshtastic
pycryptodome
